pyproject.toml
[project]
name = "solve_captcha_pje"
version = "0.0.1"

requires-python = ">= 3.11"

dependencies = [
	"numpy==1.24.1",
	"Pillow==9.5.0",
	"requests==2.28.2",
]

[tool.setuptools]
py-modules = []